feat: add event for toggling completed

+++ js/TodoList.js
@@ -9,6 +9,13 @@ const renderTodoItem = ({ id, value, completed }) => `
 `;
 
 export default function TodoList(listEl, todoApp) {
+  this.toggleCompleted = (event) => {
+    const itemEl = event.target.parentElement.parentElement;
+    const item = todoApp.getItem(itemEl.dataset.id);
+    todoApp.updateItem({ ...item, completed: !item.completed });
+    todoApp.todoInput.focus();
+  };
+
   this.deleteItem = (event) => {
     const itemEl = event.target.parentElement.parentElement;
     todoApp.deleteItem(itemEl.dataset.id);
@@ -20,6 +27,9 @@ export default function TodoList(listEl, todoApp) {
   };
 
   listEl.addEventListener("click", (event) => {
+    if (event.target.classList.contains("toggle")) {
+      return this.toggleCompleted(event);
+    }
     if (event.target.classList.contains("destroy")) {
       return this.deleteItem(event);
     }
